+++ CHANGELOG.md
@@ -14,6 +14,15 @@
   - Minor: Bugfix for `ListOption` to change class based on selected state, not selected prop only
 </details>
 
+## v0.60.0
+- feature(custom-field-input-multiple-choice): Implement `className` API
+- fix(custom-field-input-multiple-choice): Ensure read-only styled combobox has same background as the tag-list
+- feature(custom-field-input-multiple-choice): implement `placeholder` API
+- feature(custom-field-input-multiple-choice): implement `required` API
+
+## v0.59.1
+  - Patch: `undefined` coerces to `null` to avoid browser warnings when clearing Number component's input 
+
 ## v0.59.0
   - Major: `ref.value` for CustomFieldInputNumber now returns its type as Integer. To upgrade replace value with String(value)